first retry without libft

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: kazumanoda <kazumanoda@student.42.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/07/21 01:13:36 by kazumanoda        #+#    #+#             */
-/*   Updated: 2020/07/21 01:39:25 by kazumanoda       ###   ########.fr       */
+/*   Updated: 2020/07/25 17:41:45 by kazumanoda       ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,7 +32,7 @@ void			print_c(char c, t_format f);
 int				dec_len(long n);
 void			print_int(int n, t_format f);
 void			print_u(unsigned int n, t_format f);
-void			print_p(unsigned long p, t_format f);
+void			print_p(size_t p, t_format f);
 void			print_s(const char *s, t_format f);
 void			print_x(unsigned int n, t_format f);
